+++ tests/test_ipynb_to_rmd.py
@@ -7,7 +7,8 @@ from utils import list_all_notebooks, remove_outputs, remove_outputs_and_header
 @pytest.mark.parametrize('nb_file', list_all_notebooks('.ipynb'))
 def test_identity_source_write_read(nb_file):
     """
-    Test that writing the notebook with rmd, and read again, is the same as removing outputs
+    Test that writing the notebook with rmd, and read again,
+    is the same as removing outputs
     :param file:
     :return:
     """
@@ -24,7 +25,8 @@ def test_identity_source_write_read(nb_file):
 @pytest.mark.parametrize('nb_file', list_all_notebooks('.ipynb'))
 def test_identity_source_write_read_md(nb_file):
     """
-    Test that writing the notebook with rmd, and read again, is the same as removing outputs
+    Test that writing the notebook with rmd, and read again,
+    is the same as removing outputs
     :param file:
     :return:
     """